+++ src/NeatMapper/Internal/LambdaDisposable.cs
@@ -1,17 +1,24 @@
-﻿using System;
+﻿#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+#nullable disable
+#endif
+
+using System;
 
 namespace NeatMapper {
-	internal class LambdaDisposable : IDisposable {
-		private readonly Action _dispose;
+	internal sealed class LambdaDisposable : IDisposable {
+		private Action _dispose;
 
 		internal LambdaDisposable(Action dispose) {
-			_dispose = dispose;
+			_dispose = dispose
+				?? throw new ArgumentNullException(nameof(dispose));
 		}
 
 
 		private void Dispose(bool disposing) {
-			if (disposing) 
-				_dispose.Invoke();
+			if (disposing) { 
+				_dispose?.Invoke();
+				_dispose = null;
+			}
 		}
 
 		public void Dispose() {