+++ src/NeatMapper/Internal/ObjectCreationException.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 
 namespace NeatMapper {
-    internal class ObjectCreationException : InvalidOperationException{
+    internal sealed class ObjectCreationException : InvalidOperationException{
         public ObjectCreationException(Type destination, Exception exception) :
             base($"Could not create object for type {destination.Name} ({destination.FullName})", exception)
         { }