+++ src/NeatMapper/Mappers/Async/Internal/Factories/DisposableAsyncNewMapFactory.cs
@@ -20,19 +20,20 @@ namespace NeatMapper {
 
 
 		protected override void Dispose(bool disposing) {
-			lock (_disposables) { 
-				base.Dispose(disposing);
-
-				if (disposing) {
+			// We are under lock, no one can access the parent so we can safely check if if was disposed without Interlocked
+			lock (_disposables) {
+				if (disposing && _disposed != 1) {
 					foreach (var disposable in _disposables) {
 						disposable?.Dispose();
 					}
 				}
 			}
+
+			base.Dispose(disposing);
 		}
 	}
 
-	internal class DisposableAsyncNewMapFactory : DisposableAsyncNewMapFactory<object, object> {
+	internal sealed class DisposableAsyncNewMapFactory : DisposableAsyncNewMapFactory<object, object> {
 		internal DisposableAsyncNewMapFactory(Type sourceType, Type destinationType, Func<object, CancellationToken, Task<object>> mapDelegate, params IDisposable[] disposables) :
 			base(sourceType, destinationType, mapDelegate, disposables) {}
 	}