+++ src/NeatMapper/Mappers/Async/Internal/Factories/DisposableMergeMapFactory.cs
@@ -20,19 +20,20 @@ namespace NeatMapper {
 
 
 		protected override void Dispose(bool disposing) {
+			// We are under lock, no one can access the parent so we can safely check if if was disposed without Interlocked
 			lock (_disposables) {
-				base.Dispose(disposing);
-
-				if (disposing) { 
+				if (disposing && _disposed != 1) {
 					foreach (var disposable in _disposables) {
 						disposable?.Dispose();
 					}
 				}
 			}
+
+			base.Dispose(disposing);
 		}
 	}
 
-	internal class DisposableAsyncMergeMapFactory : DisposableAsyncMergeMapFactory<object, object> {
+	internal sealed class DisposableAsyncMergeMapFactory : DisposableAsyncMergeMapFactory<object, object> {
 		internal DisposableAsyncMergeMapFactory(Type sourceType, Type destinationType, Func<object, object, CancellationToken, Task<object>> mapDelegate, params IDisposable[] disposables) :
 			base(sourceType, destinationType, mapDelegate, disposables) {}
 	}