+++ src/NeatMapper/Mappers/Core/Internal/Factories/DefaultMergeMapFactory.cs
@@ -7,7 +7,7 @@ using System.Threading;
 
 namespace NeatMapper {
 	internal class DefaultMergeMapFactory<TSource, TDestination> : MergeMapFactory<TSource, TDestination> {
-		private int _disposed = 0;
+		protected int _disposed = 0;
 		private readonly Func<TSource, TDestination, TDestination> _mapDelegate;
 
 		internal DefaultMergeMapFactory(Func<TSource, TDestination, TDestination> mapDelegate)
@@ -27,16 +27,17 @@ namespace NeatMapper {
 		public override TDestination Invoke(TSource source, TDestination destination) {
 			if(Interlocked.CompareExchange(ref _disposed, 0, 0) == 1)
 				throw new ObjectDisposedException(null);
+
 			return _mapDelegate.Invoke(source, destination);
 		}
 
 		protected override void Dispose(bool disposing) {
-			if (disposing && Interlocked.CompareExchange(ref _disposed, 1, 0) == 1)
-				throw new ObjectDisposedException(null);
+			if (disposing)
+				Interlocked.CompareExchange(ref _disposed, 1, 0);
 		}
 	}
 
-	internal class DefaultMergeMapFactory : DefaultMergeMapFactory<object, object> {
+	internal sealed class DefaultMergeMapFactory : DefaultMergeMapFactory<object, object> {
 		internal DefaultMergeMapFactory(Type sourceType, Type destinationType, Func<object, object, object> mapDelegate) :
 			base(sourceType, destinationType, mapDelegate) {}
 	}