+++ src/NeatMapper/Matchers/Internal/Factories/DefaultMatchMapFactory.cs
@@ -8,7 +8,7 @@ using System.Threading;
 namespace NeatMapper {
 	internal class DefaultMatchMapFactory<TSource, TDestination> : MatchMapFactory<TSource, TDestination> {
 		protected int _disposed = 0;
-		protected readonly Func<TSource, TDestination, bool> _mapDelegate;
+		private readonly Func<TSource, TDestination, bool> _mapDelegate;
 
 		internal DefaultMatchMapFactory(Func<TSource, TDestination, bool> mapDelegate)
 			: this(typeof(TSource), typeof(TDestination), mapDelegate) { }
@@ -27,16 +27,17 @@ namespace NeatMapper {
 		public override bool Invoke(TSource source, TDestination destination) {
 			if (Interlocked.CompareExchange(ref _disposed, 0, 0) == 1)
 				throw new ObjectDisposedException(null);
+
 			return _mapDelegate.Invoke(source, destination);
 		}
 
 		protected override void Dispose(bool disposing) {
-			if (disposing && Interlocked.CompareExchange(ref _disposed, 1, 0) == 1)
-				throw new ObjectDisposedException(null);
+			if (disposing)
+				Interlocked.CompareExchange(ref _disposed, 1, 0);
 		}
 	}
 
-	internal class DefaultMatchMapFactory : DefaultMatchMapFactory<object, object> {
+	internal sealed class DefaultMatchMapFactory : DefaultMatchMapFactory<object, object> {
 		internal DefaultMatchMapFactory(Type sourceType, Type destinationType, Func<object, object, bool> mapDelegate) :
 			base(sourceType, destinationType, mapDelegate) { }
 	}