+++ src/NeatMapper/Options/Configuration/Delegates/MatchMapDelegate.cs
@@ -1,13 +1,13 @@
 ﻿namespace NeatMapper {
 	/// <summary>
-	/// Delegate which allows matching two objects of two given types, used to override <see cref="IMatchMap{TSource, TDestination}"/>
+	/// Delegate which allows matching two objects of two given types, used to override <see cref="IMatchMap{TSource, TDestination}"/>.
 	/// </summary>
-	/// <typeparam name="TSource">Source type</typeparam>
-	/// <typeparam name="TDestination">Destination type</typeparam>
-	/// <param name="source">Source object, may be null</param>
-	/// <param name="destination">Destination object, may be null</param>
+	/// <typeparam name="TSource">Source type.</typeparam>
+	/// <typeparam name="TDestination">Destination type.</typeparam>
+	/// <param name="source">Source object, may be null.</param>
+	/// <param name="destination">Destination object, may be null.</param>
 	/// <param name="context">Matching context, which allows nested matches, services retrieval via DI, ...</param>
-	/// <returns><see langword="true"/> if the two objects match</returns>
+	/// <returns><see langword="true"/> if the two objects match.</returns>
 	public delegate bool MatchMapDelegate<TSource, TDestination>(
 #if NET5_0_OR_GREATER
 		TSource?